requirements.txt
apache-airflow-providers-google==10.13.1
dbt-bigquery==1.7.4
dbt-core==1.7.4
